+++ src/views/PluginEdit.vue
@@ -1,29 +1,102 @@
 <template>
   <div class="p4">
+    <portal
+      to="app-bar-extension"
+    >
+      <div style="display: flex; flex-direction: row; width:100%">
+        <v-tabs
+          v-model="tab"
+          align-with-title
+        >
+          <v-tab
+            key="interact"
+            href="#interact"
+          >
+            Interact
+            <v-icon x-small class="ml-1">
+              fa-terminal
+            </v-icon>
+          </v-tab>
+          <v-tab
+            key="tasks"
+            href="#tasks"
+          >
+            Tasks
+            <v-icon x-small class="ml-1">
+              fa-sticky-note
+            </v-icon>
+          </v-tab>
+        </v-tabs>
+      </div>
+    </portal>
     <edit-page-top
       :breads="breads"
-      :show-submit="true"
+      :show-submit="false"
       :show-copy="false"
       :show-delete="false"
       :submit-loading="loading"
       @submit="submit"
+    >
+      <template slot="extra-stuff">
+        <tooltip-button-toggle
+          v-model="isRefreshTasks"
+          icon="fa-redo"
+          :button-text="isRefreshTasks ? 'On' : 'Off'"
+          text="Auto-refresh tasks"
+        />
+      </template>
+    </edit-page-top>
+    <error-state-alert
+      v-if="errorState"
+      :resource-id="id"
+      resource-type="plugin"
     />
-    <h4 class="pl-4 pb-4">
-      Execute Plugin
-    </h4>
-    <v-card style="padding: 10px">
-      <info-viewer
-        class="info-viewer"
-        :info="pluginInfo"
-      />
-      <technique-chips :techniques="plugin.TechniqueChips" />
-      <general-form
-        v-if="reset"
-        ref="generalform"
-        v-model="form"
-        :options="plugin.options"
-      />
-    </v-card>
+    <v-tabs-items
+      v-else
+      v-model="tab"
+      class="scrollable-pane"
+    >
+      <v-tab-item
+        key="interact"
+        :value="'interact'"
+        :transition="false"
+        :reverse-transition="false"
+      >
+        <h4 class="pl-4 pb-4">
+          Execute Plugin
+        </h4>
+        <v-card style="padding: 10px">
+          <info-viewer
+            class="info-viewer"
+            :info="pluginInfo"
+          />
+          <technique-chips :techniques="plugin.TechniqueChips" />
+          <general-form
+            v-if="reset"
+            ref="generalform"
+            v-model="form"
+            :options="plugin.options"
+          />
+          <v-btn
+            :loading="loading"
+            color="primary"
+            @click="submit"
+          >
+            Submit
+          </v-btn>
+        </v-card>
+      </v-tab-item>
+      <v-tab-item
+        key="tasks"
+        :value="'tasks'"
+        :transition="false"
+        :reverse-transition="false"
+      >
+        <v-card flat>
+          <plugin-tasks-list :plugin="plugin" :refresh-tasks="isRefreshTasks" />
+        </v-card>
+      </v-tab-item>
+    </v-tabs-items>
   </div>
 </template>
 
@@ -33,6 +106,9 @@ import InfoViewer from '@/components/InfoViewer.vue';
 import * as pluginApi from '@/api/plugin-api';
 import TechniqueChips from '@/components/TechniqueChips.vue';
 import EditPageTop from '@/components/EditPageTop.vue';
+import PluginTasksList from '@/components/plugins/PluginTasksList.vue';
+import TooltipButtonToggle from '@/components/TooltipButtonToggle.vue';
+import ErrorStateAlert from '@/components/ErrorStateAlert.vue';
 
 export default {
   name: 'PluginEdit',
@@ -41,13 +117,18 @@ export default {
     GeneralForm,
     TechniqueChips,
     EditPageTop,
+    PluginTasksList,
+    TooltipButtonToggle,
+    ErrorStateAlert,
   },
   data() {
     return {
       reset: true,
       loading: false,
+      isRefreshTasks: false,
       form: {},
-      plugin: {},
+      plugin: { options: {} },
+      errorState: false,
     };
   },
   computed: {
@@ -88,6 +169,14 @@ export default {
     id() {
       return this.$route.params.id;
     },
+    tab: {
+      set(tab) {
+        this.$router.replace({ query: { ...this.$route.query, tab } });
+      },
+      get() {
+        return this.$route.query.tab || 'interact';
+      },
+    },
   },
   mounted() {
     this.getPlugin(this.id);
@@ -125,5 +214,12 @@ export default {
 };
 </script>
 
-<style>
+<style lang="scss">
+// Overrides vuetify.css
+// Because we moved the tabs into a div, which made the color funky.
+.v-toolbar__content > div > .v-tabs > .v-slide-group.v-tabs-bar,
+.v-toolbar__extension > div > .v-tabs > .v-slide-group.v-tabs-bar {
+  background-color: inherit;
+}
+
 </style>